# streaming/demos/streaming/document-ocr-streaming/source/openapi.json
{"openapi":"3.0.2","info":{"title":"Gradio","description":"","version":"6.20.0"},"paths":{"/gradio_api/upload":{"post":{"summary":"Upload File","operationId":"upload_file_upload_post","parameters":[{"name":"upload_id","in":"query","required":false,"schema":{"type":"string","nullable":true,"default":null},"description":"Optional ID to track upload progress"}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"files":{"type":"array","items":{"type":"string","format":"binary"},"description":"One or more files to upload"}},"required":["files"]}}}},"responses":{"200":{"description":"List of file paths where the uploaded files were saved","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"400":{"description":"Invalid content type or invalid file name","content":{"text/plain":{"schema":{"type":"string"}}}},"413":{"description":"File exceeds maximum allowed size","content":{"text/plain":{"schema":{"type":"string"}}}}},"security":[{"login_check":[]}]}},"/gradio_api/call/v2/run_ocr":{"post":{"summary":"Stream a bounded group of pages within one ZeroGPU reservation. Every page is rasterized and inferred independently and sequentially. The bounded group amortizes ZeroGPU scheduling while keeping long PDFs split across multiple leases.. File inputs must first be uploaded via POST /gradio_api/upload (multipart/form-data with a \"files\" field). Use the returned path in the request body as {\"path\": \"<uploaded_path>\", \"meta\": {\"_type\": \"gradio.FileData\"}}.","description":"Stream a bounded group of pages within one ZeroGPU reservation. Every page is rasterized and inferred independently and sequentially. The bounded group amortizes ZeroGPU scheduling while keeping long PDFs split across multiple leases.Output must be fetched from GET /gradio_api/call/run_ocr/{event_id}. Returns an array of 1 elements of the following format: ['output ({})']","operationId":"run_ocr","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"image_path":{"type":"string","description":"Must include {\"path\": \"url\", \"meta\": {\"_type\": \"gradio.FileData\"}}. The meta key signals that the file will be uploaded to the remote server.","format":"filepath"},"page_index":{"type":"integer"},"page_count":{"type":"integer"}}}}}},"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"type":"object","properties":{"event_id":{"type":"string"}}}}}}}}},"/gradio_api/call/run_ocr/{event_id}":{"get":{"summary":"Fetch results for /run_ocr","description":"Returns a stream of server-sent events (SSE). The final event has `event: complete` with `data` containing a JSON array of outputs.","operationId":"run_ocr_get","parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"string"},"description":"The event_id returned by the POST request"}],"responses":{"200":{"description":"SSE stream with event: complete containing a JSON array of outputs","content":{"text/event-stream":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{}}}